pyproject.toml
[tool.poetry]
name = "dealnest-api"
version = "0.1.0"
description = ""
authors = ["Emeka Rex Chukwu <emekarexchukwu@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.116.2"
python-dotenv = "^1.1.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic-settings = "^2.10.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
email-validator = "^2.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
flake8 = "^7.3.0"
black = "^25.9.0"
isort = "^6.0.1"
httpx = "^0.27.0"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
